backend/requirements.txt
# Conda
cryptography==40.0.2
fastapi==0.95.1
gunicorn==20.1.0
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
passlib==1.7.4
pydantic==1.10.7
pymysql==1.1.0
scikit-image==0.20.0
scikit-learn==1.2.2
SQLAlchemy==2.0.9
starlette==0.26.1
uvicorn==0.21.1

# Pip
lime==0.2.0.1
plspm==0.5.6
python-jose==3.3.0
python-multipart==0.0.6
python-ulid==1.1.0
SALib==1.4.7
python-dotenv==1.0.0
